extract network notifier
